requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0 
anthropic==0.*
python-dotenv==1.0.1
tokenizers
sqlalchemy[asyncio]==2.*
aiosqlite==0.19.0
alembic==1.12.1
asyncpg==0.29.0
psycopg2-binary==2.9.9
feedparser==6.*
newspaper3k==0.2.*
lxml[html_clean]==5.2.2
sentence-transformers==3.0.1
transformers==4.41.2
huggingface-hub
numpy==1.24.*
hdbscan==0.8.*
scikit-learn==1.3.*
snscrape
requests
urllib3
praw
youtube-transcript-api
google-api-python-client 